requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=1.0.0=pypi_0
astunparse=1.6.3=pypi_0
blas=1.0=mkl
bottleneck=1.3.2=py37heb32a55_1
brotli=1.0.9=he6710b0_2
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.26=h06a4308_2
cached-property=1.5.2=pypi_0
cachetools=4.2.4=pypi_0
certifi=2021.10.8=py37h06a4308_2
charset-normalizer=2.0.10=pypi_0
click=8.0.3=pyhd3eb1b0_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
expat=2.4.1=h2531618_2
ffmpeg=4.2.2=h20bf706_0
flatbuffers=2.0=pypi_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
gast=0.4.0=pypi_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h5202010_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.3.3=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.43.0=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=3.6.0=pypi_0
icu=58.2=he6710b0_3
idna=3.3=pypi_0
imageio=2.9.0=pyhd3eb1b0_0
importlib-metadata=4.10.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
keras=2.7.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=py37h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libclang=12.0.0=pypi_0
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libidn2=2.3.2=h7f8727e_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.6=pypi_0
matplotlib=3.5.0=py37h06a4308_0
matplotlib-base=3.5.0=py37h3ed280b_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
munkres=1.1.4=py_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
nltk=3.6.5=pyhd3eb1b0_0
numexpr=2.8.1=py37h6abb31d_0
numpy=1.21.2=py37h20f2e39_0
numpy-base=1.21.2=py37h79a1101_0
oauthlib=3.1.1=pypi_0
olefile=0.46=py37_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1m=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py37h8c16a72_0
pcre=8.45=h295c915_0
pillow=8.4.0=py37h5aabda8_0
pip=21.2.2=py37h06a4308_0
protobuf=3.19.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pydot=1.4.2=pypi_0
pydotplus=2.0.2=pypi_0
pyparsing=3.0.6=pypi_0
pyqt=5.9.2=py37h05f1152_2
python=3.7.11=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.10.1=py3.7_cuda10.2_cudnn7.6.5_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
regex=2021.8.3=py37h7f8727e_0
requests=2.27.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.8=pypi_0
setuptools=58.0.4=py37h06a4308_0
sip=4.19.8=py37hf484d3e_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.37.0=hc218d9a_0
tensorboard=2.7.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow-estimator=2.7.0=pypi_0
tensorflow-gpu=2.7.0=pypi_0
tensorflow-io-gcs-filesystem=0.23.1=pypi_0
termcolor=1.1.0=pypi_0
tk=8.6.11=h1ccaba5_0
torchaudio=0.10.1=py37_cu102
torchvision=0.11.2=py37_cu102
tornado=6.1=py37h27cfd23_0
tqdm=4.62.3=pyhd3eb1b0_1
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.8=pypi_0
werkzeug=2.0.2=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
wrapt=1.13.3=pypi_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.5=h7b6447c_0
zipp=3.7.0=pypi_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
